fix error data type post merge

+++ src/main/java/com/vias/uc/backend/model/Usuario.java
@@ -15,7 +15,7 @@ public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_usuario")
-    private Long idUsuario;
+    private Integer idUsuario;
 
     @Column(nullable = false)
     private String nombre;
@@ -72,12 +72,12 @@ public class Usuario {
 
     // ======= getters / setters =======
 
-    public Long getIdUsuario() {
-        return idUsuario.longValue();
+    public Integer getIdUsuario() {
+        return idUsuario;
     }
 
     public void setIdUsuario(Integer idUsuario) {
-        this.idUsuario = idUsuario.longValue();
+        this.idUsuario = idUsuario;
     }
 
     public String getNombre() {
